extracted duplicated localization test helpers to a separate file
