+++ EssentialFeedTests/Feed Presentation/FeedPresenterTests.swift
@@ -17,7 +17,7 @@ class FeedPresenterTests: XCTestCase {
         XCTAssertEqual(FeedPresenter.title, expectedTitle)
     }
 
-    func test_loadError_hasLocalizedError() {
+    func test_loadError_hasLocalizedGenericError() {
         let expectedKey = "GENERIC_CONNECTION_ERROR"
         let expectedTitle = localized(key: expectedKey, in: "Shared")
 
@@ -67,56 +67,13 @@ class FeedPresenterTests: XCTestCase {
     }
 
     func test_localizedStrings_haveKeysAndValuesForAllSupportedLocalizations() {
-            let table = "Feed"
-            let presentationBundle = Bundle(for: FeedPresenter.self)
-            let localizationBundles = allLocalizationBundles(in: presentationBundle)
-            let localizedStringKeys = allLocalizedStringKeys(in: localizationBundles, table: table)
+        let table = "Feed"
+        let bundle = Bundle(for: FeedPresenter.self)
 
-            localizationBundles.forEach { (bundle, localization) in
-                localizedStringKeys.forEach { key in
-                    let localizedString = bundle.localizedString(forKey: key, value: nil, table: table)
-
-                    if localizedString == key {
-                        let language = Locale.current.localizedString(forLanguageCode: localization) ?? ""
-
-                        XCTFail("Missing \(language) (\(localization)) localized string for key: '\(key)' in table: '\(table)'")
-                    }
-                }
-            }
-        }
+        assertStringsLocalized(for: bundle, in: table)
+    }
 
     // MARK: - Helpers
-
-    private typealias LocalizedBundle = (bundle: Bundle, localization: String)
-
-    private func allLocalizationBundles(in bundle: Bundle, file: StaticString = #file, line: UInt = #line) -> [LocalizedBundle] {
-        return bundle.localizations.compactMap { localization in
-            guard
-                let path = bundle.path(forResource: localization, ofType: "lproj"),
-                let localizedBundle = Bundle(path: path)
-            else {
-                XCTFail("Couldn't find bundle for localization: \(localization)", file: file, line: line)
-                return nil
-            }
-
-            return (localizedBundle, localization)
-        }
-    }
-
-    private func allLocalizedStringKeys(in bundles: [LocalizedBundle], table: String, file: StaticString = #file, line: UInt = #line) -> Set<String> {
-        return bundles.reduce([]) { (acc, current) in
-            guard
-                let path = current.bundle.path(forResource: table, ofType: "strings"),
-                let strings = NSDictionary(contentsOfFile: path),
-                let keys = strings.allKeys as? [String]
-            else {
-                XCTFail("Couldn't load localized strings for localization: \(current.localization)", file: file, line: line)
-                return acc
-            }
-
-            return acc.union(Set(keys))
-        }
-    }
 
     private class FeedViewSpy: FeedLoadingView, FeedView, FeedErrorView {
         enum Message: Hashable {
